pyproject.toml
[tool.poetry]
name = "lichclient"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
httpx = "^0.27.2"
authlib = "^1.3.2"
python-dotenv = "^1.0.1"
pydantic = "^2.8.2"


[tool.poetry.scripts]
client = "lichclient.main:run"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
